backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow>=10.2.0
numpy>=1.26.0
pandas>=2.1.3
scikit-learn>=1.3.2
tensorflow>=2.14.0
python-multipart==0.0.6